USP-MoE/requirements.txt
accelerate==0.23.0
einops==0.7.0
ema_pytorch==0.2.3
matplotlib==3.5.3
numpy==1.23.2
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.1.2
scipy==1.9.1
torch_geometric==2.2.0
torchsummary==1.5.1
tqdm==4.64.0
xlrd==2.0.1
xlwt==1.3.0
`1